New Feature: Option to allow dues to be paid ahead
Adds a configuration option "Max dues year that can be paid" which
defaults to the year you install the plugin (or when you upgrade to the
version that contains it).  It's ignored if the member isn't current,
but if they are current, they'll still get a link to pay their dues if
their dues year is less than the one you put in this option. This can be
used to offer to pay the following year's dues if you allow them (or
require them) to be paid before the year starts.